feat: Persistent Manager 구현
